boards/stm32f429i-disco: add support
The boards/stm32f429i-disco board is an ealier version of the stm32f429i-disc1.

The boards are compatible with the difference that the older on-board debugger
does not provide a UART connection.

To spare the user of having to wire up a USB-TTL adapter, enable CDC-ACM
on the micro-USB port.