tests/posix_semaphore: better guarantee sequence of output lines
As suggested by @kaspar030 the THREAD_CREATE_WOUT_YIELD flag was removed.
Now the sequence of output lines is identical in two cases: stdio via UART,
and stdio via CDC ACM (which needs a thread for that).

The expects have been changed to match the new situation, which is what
was already given in the CDC ACM case.